% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProxSeek_runProxeQTL.R
\name{ProxSeek_runProxeQTL}
\alias{ProxSeek_runProxeQTL}
\title{Run Proximity-based eQTL Analysis}
\usage{
ProxSeek_runProxeQTL(SNP_file_name, expression_file_name, snps_location_file_name,
       gene_location_file_name, covariates_file_name, cell.type = NA, chr = NA,
       output.file = "./test/ciseqtl.tsv", n.cluster = 4)
}
\arguments{
\item{SNP_file_name}{Path to the SNP genotype file.}

\item{expression_file_name}{Path to the gene expression file.}

\item{snps_location_file_name}{Path to the file with SNP locations.}

\item{gene_location_file_name}{Path to the file with gene locations.}

\item{covariates_file_name}{Path to the covariates file.}

\item{cell.type}{Specific cell type to analyze.}

\item{chr}{integer. Chromosome to analyze.}

\item{output.file}{Output file path for the results.}

\item{n.cluster}{Number of clusters for analysis.}

\item{P_PIR_folder}{The folder where P-PIR dataset were downloaded.}
}
\value{
A data frame with the results of the proximity-based eQTL analysis.
}
\description{
This function performs proximity-based eQTL analysis using SNP and gene expression data,
integrating cell type-specific physical interaction regions (P-PIRs). It outputs significant
Prox-eQTLs, considering cellular context.
}
\details{
This function integrates MatrixeQTL pipeline and P-PIR information to identify prox-eQTLs.
}
\examples{
# Example usage of ProxSeek_runProxeQTL
\dontrun{
result <- ProxSeek_runProxeQTL("path/to/SNP_file.txt", "path/to/expression_file.txt",
                               "path/to/snps_location_file.txt", "path/to/gene_location_file.txt",
                               "path/to/covariates_file.txt", cell.type = "monocyte", chr = "X")
}
}
