% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProxSeek_selectProxCis.R
\name{ProxSeek_selectProxCis}
\alias{ProxSeek_selectProxCis}
\title{ProxSeek_selectProxCis}
\usage{
ProxSeek_selectProxCis(
  ciseqtl.file = NA,
  P_PIR_folder = NA,
  cell.type = NA,
  n.cluster = 4
)
}
\arguments{
\item{ciseqtl.file}{A string specifying the path to the cis-eQTL file.
Default value is NA. It's recommended to provide the actual file path for specific analyses.}

\item{cell.type}{A string specifying the cell type. This is used to construct the file name of the cell type-specific P-PIR file.}

\item{n.cluster}{An integer specifying the number of clusters to use in parallel processing. Default is 4.
This should be set according to the hardware capabilities of the system.}
}
\value{
An annotated data frame with the results of the cis-eQTL and Hi-C interaction processing.
}
\description{
This function select cis-eQTLs that lie within promoter and promoter interaction region (P-PIR) based on tissue/cell type Hi-C data.
It reads cis-eQTL and cell type P-PIR files, performs filtering and annotation,
and returns an annotated data frame.
}
\examples{
# Example usage:
\dontrun{
demo.ciseqtl.file<-system.file("data", "ciseqtl.csv", package = "ProxSeek")
result <- ProxSeek_selectProxCis(ciseqtl.file = demo.ciseqtl.file, P_PIR_folder = /path/to/downloaded/PIRfiles/,
cell.type = "monocyte", n.cluster =4)
}


}
